discordplays/requirements.txt
discord.py
python-dotenv
pydirectinputs